+++ packages/navigation/src/header-height-context/headerHeightContext.tsx
@@ -30,16 +30,8 @@ export const HeaderHeightProvider = ({ children }: PropsWithChildren) => {
 };
 
 /**
- * Used within mad navigation to provide the header height
- */
-const _useMadNavigationHeaderHeight = () => {
-    const context = useContext(HeaderHeightContext);
-    return context.setHeight;
-};
-
-/**
  * Get the header height
- * @returns a number
+ * @returns {number} header height
  */
 export const useHeaderHeight = () => {
     const context = useContext(HeaderHeightContext);
@@ -51,7 +43,7 @@ export const useHeaderHeight = () => {
  */
 export const HeightSender = () => {
     const headerHeight = useReactNavigationHeaderHeight();
-    const setHeight = _useMadNavigationHeaderHeight();
+    const { setHeight } = useContext(HeaderHeightContext);
     useEffect(() => {
         setHeight?.(() => headerHeight);
         // eslint-disable-next-line react-hooks/exhaustive-deps -- we only want this to run when header height changes